BREAKING: remove Portal wrapper from components
Currently we wrap certain components such as FAB.Group and Snackbar in Portal by default.
This means they are always rendered on top of everything and there's no way to control it.
This makes it hard to integrate with navigation libraries such as React Navigation.
The PR removes the Portal wrapper by default and exports PortalHost so users can control where the components are rendered.

In future React Navigation could provide an API to render things on top of the screen and users would be able to take advantage of it.