Allow TouchableRipple for Modern Android Versions
TouchableRipple uses hard-coded values to prohibit use on Android versions > 27. This means that each time a new version of android is released, the maintainers of this library must update this file to allow TouchableRipple to be used. The proposed change is to allow TouchableRipple to be used on newer versions of Android, namely API level 28.