feat: add  style for internal DataTableRow view
The current implementation enforces a paddingHorizontal of 16 as default for the content style. This style is inherited by both TouchableRipple and its child view. TouchableRipple container values can be overridden by passing the style property value when creating a DataTableRow, but the child view will always have the paddingHorizontal value of 16. As part of this commit we will allow the style class property to override the child view as well.