fix: use route keys for lazy loading tabs instead of index
Previously, we kept an array with indices of the routes we've visited previously. However, if routes were re-ordered/removed, the indices may not represent the actual routes that were visited previously since the routes might have a different index after re-ordering.

Using route keys instead of indices makes sure that we only render screens that were visited previously even when routes are re-ordered.