refactor: more logical calculation of TextInputFlat label position
The padding used to calculate the label's topPosition was previously
`minInputHeight / 2` when the input has no explicit height. While this
happened to produce acceptable behaviour with the current constant
values, the input height is not the logical basis for the padding, but
rather the additional height of the TextInputFlat that is reserved for
the minimised label. With the new calculation, the constants can be
changed (in particular MINIMIZED_LABEL_LOGICAL_MARGIN_TOP) and the
component will still have a correct layout.

Note that I have added 2 pixels to that padding in on top of what one
might expect from the constants. This is to preserve the same layout as
before when using the current constants. I think it would make sense to
remove this fudge factor (and it still looks good), but I have included
it in the interest of trying not to disturb any pixels in this commit.