feat: customize top padding of a flat TextInput
A `paddingTop` value in a TextInput's style now controls the spacing
between the top of a flat TextInput and the top of its minimized label,
which defaults to 12.

Previously this was hardcoded to 12, even when setting dense to true,
which makes it difficult to style flat TextInputs so that they take up
minimal vertical space. Dense could be used and a small height given,
but 12 pixels would always be frustratingly reserved above the minimized
label. Now the paddingTop can be set to smaller value to get a more
pleasing vertically compact view.

The use of the style's paddingTop feels appropriate given the effect of
value and the use of other style properties to control the internal
layout of TextInputFlat, but I'm open to suggestions of alternative
approaches to exposing this.

Partially addresses #2700.