fix: datatable.title cell height multiple of numberOfLines
Without this, it does not appear the numberOfLines prop has an effect, as noticed in #848 and #863
With this, it appears to work exactly as expected.

Text components require special alignment treatment with word wrapping. Normally a word wrap
will force text alignment to flex-start. So we align directly when number of lines is > 1, while
respecting the numeric property and it's interaction with RTL languages