feat: allow menu anchor position to be explicitly set
Although Menu.tsx comments claim that menus typically open at the bottom of their anchors, the code itself seems to mostly keep `top` (the top of the menu) the same as `anchorLayout.y`, which means the menu always opens _over_ the anchor.

This change suggests adding an optional `anchorPosition` property that, when set to `"bottom"`, will instead put the menu at the bottom of its anchor.