fix(docs): `SegementedButton` theme docs
change docs to accurately reflect the color used for the
`backgroundColor` of the `SegementedButton` component when it is
unchecked.

Code for the `backgroundColor` is [here](https://github.com/callstack/react-native-paper/blob/015e418c600b8c39334d65138decb46787def3c7/src/components/SegmentedButtons/utils.ts#L87C12-L87C12)
and it clearly returns `transparent` when the button is not checked.