Update 01-getting-started.md, add a note for using expo-react-native-paper
This commit introduces a link to the `expo-react-native-paper` template in the `README.md` file. The `expo-react-native-paper` template is a valuable resource for developers starting new projects with React Native Paper and Expo.

The template is designed with pre-configured navigation examples using Expo's router, which can significantly expedite the development process. By including this link, we aim to provide developers with a quick and efficient way to kickstart their React Native projects.

We believe that this addition will enhance the usability of `react-native-paper` and provide a more comprehensive experience for developers.