fix: bottom navigation shifting={false} active pill display
When the bottom nav has the attribute shifting={false} the pill shape is
misformed on the active tabs. On clicking around this seems
inconsistent. (See https://github.com/callstack/react-native-paper/issues/4819)

This appears to be because the animated view is added / removed as the
tab is focused causing the animation to not play or half play. Since it
starts at 0.5 scaleX this appears to the user as a squashed pill shape.
The intended animation is to expand out from the center.

This commit fixes this by ensuring the animation is always present and
simply varies the outputRange of the animation depending on the focused
state.